requirements.txt
absl-py==1.4.0
anticaptchaofficial==1.0.56
asgiref==3.7.2
beautifulsoup4==4.12.2
blinker==1.6.2
boto3==1.28.27
botocore==1.31.27
bs4==0.0.1
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.6
cmake==3.27.4.1
colorama==0.4.6
contourpy==1.1.0
cryptography==41.0.3
cycler==0.10.0
distlib==0.3.7
Django==4.2.5
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
dnspython==2.4.2
docx2pdf==0.1.8
easyocr==1.7.0
et-xmlfile==1.1.0
filelock==3.12.2
fire==0.5.0
Flask==2.3.2
Flask-Cors==4.0.0
fonttools==4.42.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
huggingface-hub==0.16.4
idna==2.10
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.4
lazy_loader==0.3
lit==16.0.6
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
ninja==1.11.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
platformdirs==3.10.0
protobuf==4.24.0
psutil==5.9.5
py==1.11.0
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybboxes==0.1.6
pyclipper==1.3.0.post4
pycparser==2.21
PyJWT==1.7.1
pymongo==4.5.0
pyparsing==2.4.7
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
roboflow==1.1.2
rsa==4.9
s3transfer==0.6.2
sahi==0.11.14
scikit-image==0.21.0
scipy==1.11.1
seaborn==0.12.2
shapely==2.0.1
six==1.16.0
smmap==5.0.0
soupsieve==2.5
sqlparse==0.4.4
supervision==0.13.0
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
tifffile==2023.8.12
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
ultralytics==8.0.153
urllib3==1.26.16
virtualenv==20.24.3
Werkzeug==2.3.6
wget==3.2
yolov5==7.0.12
zipp==3.16.2
